feat: 增加jwt配置减少查库，提高响应速度
